// src/types/contracts.ts
export interface CreateTokenParams {
  credentialId: string;
  name: string;
  symbol: string;
  emissionRate: number;
  maxSupply: number;
  initialLiquidity: {
    tokenAmount: number;
    ethAmount: number;
  };
}